requirements.txt
astroid==2.3.2
attrs==19.3.0
bcrypt==3.1.7
certifi==2019.9.11
cffi==1.13.1
chardet==3.0.4
Click==7.0
colorama==0.4.1
Flask==1.1.1
flask-mongoengine==0.9.5
Flask-Script==2.0.6
Flask-WTF==0.14.2
idna==2.8
importlib-metadata==0.23
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.3
jsonschema==3.1.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mongoengine==0.18.2
more-itertools==7.2.0
pycparser==2.19
pycryptodome==3.9.4
pylint==2.4.3
pymongo==3.9.0
pyrsistent==0.15.4
requests==2.22.0
six==1.12.0
typed-ast==1.4.0
urllib3==1.25.7
Werkzeug==0.16.0
wrapt==1.11.2
WTForms==2.2.1
zipp==0.6.0
